#434: Moved the PassiveSkillModifierIncreaseParser to the model folder
 The file was not changed yet.
